requirements.txt
fastapi==0.100.0
  pydantic==2.0.2
    annotated-types==0.5.0
    pydantic_core==2.1.2
      typing_extensions==4.7.1
    typing_extensions==4.7.1
  starlette==0.27.0
    anyio==3.7.1
      exceptiongroup==1.1.2
      idna==3.4
      sniffio==1.3.0
  typing_extensions==4.7.1
httptools==0.6.0
ipython==8.14.0
  backcall==0.2.0
  decorator==5.1.1
  jedi==0.18.2
    parso==0.8.3
  matplotlib-inline==0.1.6
    traitlets==5.9.0
  pexpect==4.8.0
    ptyprocess==0.7.0
  pickleshare==0.7.5
  prompt-toolkit==3.0.39
    wcwidth==0.2.6
  Pygments==2.15.1
  stack-data==0.6.2
    asttokens==2.2.1
      six==1.16.0
    executing==1.2.0
    pure-eval==0.2.2
  traitlets==5.9.0
manga-ocr==0.1.10
  fire==0.5.0
    six==1.16.0
    termcolor==2.3.0
  fugashi==1.2.1
  jaconv==0.3.4
  loguru==0.7.0
  numpy==1.25.1
  Pillow==10.0.0
  pyperclip==1.8.2
  torch==2.0.1
    filelock==3.12.2
    Jinja2==3.1.2
      MarkupSafe==2.1.3
    networkx==3.1
    nvidia-cublas-cu11==11.10.3.66
      setuptools==65.5.0
      wheel==0.40.0
    nvidia-cuda-cupti-cu11==11.7.101
      setuptools==65.5.0
      wheel==0.40.0
    nvidia-cuda-nvrtc-cu11==11.7.99
    nvidia-cuda-runtime-cu11==11.7.99
      setuptools==65.5.0
      wheel==0.40.0
    nvidia-cudnn-cu11==8.5.0.96
      nvidia-cublas-cu11==11.10.3.66
        setuptools==65.5.0
        wheel==0.40.0
    nvidia-cufft-cu11==10.9.0.58
    nvidia-curand-cu11==10.2.10.91
      setuptools==65.5.0
      wheel==0.40.0
    nvidia-cusolver-cu11==11.4.0.1
      nvidia-cublas-cu11==11.10.3.66
        setuptools==65.5.0
        wheel==0.40.0
    nvidia-cusparse-cu11==11.7.4.91
      setuptools==65.5.0
      wheel==0.40.0
    nvidia-nccl-cu11==2.14.3
    nvidia-nvtx-cu11==11.7.91
      setuptools==65.5.0
      wheel==0.40.0
    sympy==1.12
      mpmath==1.3.0
    triton==2.0.0
      cmake==3.26.4
      filelock==3.12.2
      lit==16.0.6
    typing_extensions==4.7.1
  transformers==4.30.2
    filelock==3.12.2
    huggingface-hub==0.16.4
      filelock==3.12.2
      fsspec==2023.6.0
      packaging==23.1
      PyYAML==6.0
      requests==2.31.0
        certifi==2023.5.7
        charset-normalizer==3.2.0
        idna==3.4
        urllib3==2.0.3
      tqdm==4.65.0
      typing_extensions==4.7.1
    numpy==1.25.1
    packaging==23.1
    PyYAML==6.0
    regex==2023.6.3
    requests==2.31.0
      certifi==2023.5.7
      charset-normalizer==3.2.0
      idna==3.4
      urllib3==2.0.3
    safetensors==0.3.1
    tokenizers==0.13.3
    tqdm==4.65.0
  unidic-lite==1.0.8
pip==23.1.2
pipdeptree==2.9.5
python-dotenv==1.0.0
python-multipart==0.0.6
uvicorn==0.22.0
  click==8.1.4
  h11==0.14.0
uvloop==0.17.0
watchfiles==0.19.0
  anyio==3.7.1
    exceptiongroup==1.1.2
    idna==3.4
    sniffio==1.3.0
websockets==11.0.3
